refactor1: remove virtual solver class, solver returns SimulationData instead of OdeSolution
